requirements.txt
pyyaml>=6.0
jsonschema>=4.23.0
pytest>=8.0.0
